start linting
