+++ src/components/AngerNoodle/helperFunctions/eventHandlers.js
@@ -1,11 +1,15 @@
-export const handleKeyPress = (event, gameState, updateGameState, snake, updateSnake, updateSnakePosition) => {
-    if (event.key === ' ') {
-        if (gameState === "active"){
-            updateGameState("paused");
-        } else {
-            updateGameState("active");
-            updateSnake(updateSnakePosition(snake));
-        }
-        console.log('space press here! ');
+const handleKeyPress = (
+  event, gameState, updateGameState, snake, updateSnake, updateSnakePosition,
+) => {
+  if (event.key === ' ') {
+    if (gameState === 'active') {
+      updateGameState('paused');
+    } else {
+      updateGameState('active');
+      updateSnake(updateSnakePosition(snake));
     }
-}
+    // console.log('space press here! ');
+  }
+};
+
+export default handleKeyPress;
